feat: ArchUnit enforces rules for other classes
