+++ src/main/java/dev/abarmin/bots/controller/TelegramBotController.java
@@ -6,7 +6,7 @@ import com.pengrad.telegrambot.model.Update;
 import com.pengrad.telegrambot.request.SetWebhook;
 import dev.abarmin.bots.service.TelegramUpdateService;
 import dev.abarmin.bots.config.DigestBotProperties;
-import dev.abarmin.bots.rss.digest.DigestBotUpdate;
+import dev.abarmin.bots.model.DigestBotUpdate;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.boot.ApplicationArguments;
