+++ src/main/java/dev/abarmin/bots/repository/TelegramBotUpdateRepository.java
@@ -1,6 +1,6 @@
 package dev.abarmin.bots.repository;
 
-import dev.abarmin.bots.rss.persistence.TelegramBotUpdate;
+import dev.abarmin.bots.entity.TelegramBotUpdate;
 import org.springframework.data.repository.ListCrudRepository;
 
 public interface TelegramBotUpdateRepository extends ListCrudRepository<TelegramBotUpdate, Integer> {
