+++ src/main/java/dev/abarmin/bots/service/ArticleSourceService.java
@@ -1,6 +1,6 @@
 package dev.abarmin.bots.service;
 
-import dev.abarmin.bots.rss.reader.persistence.ArticleSource;
+import dev.abarmin.bots.entity.ArticleSource;
 
 import java.net.URI;
 import java.util.Collection;
