+++ src/main/java/dev/abarmin/bots/service/digest/CreatedStateHandler.java
@@ -1,18 +1,18 @@
-package dev.abarmin.bots.rss.digest.operation;
+package dev.abarmin.bots.service.digest;
 
 import com.pengrad.telegrambot.TelegramBot;
 import com.pengrad.telegrambot.model.Update;
 import com.pengrad.telegrambot.model.request.KeyboardButton;
 import com.pengrad.telegrambot.model.request.ReplyKeyboardMarkup;
 import com.pengrad.telegrambot.request.SendMessage;
-import dev.abarmin.bots.core.BotHelper;
-import dev.abarmin.bots.core.BotOperation;
-import dev.abarmin.bots.core.MessageSourceHelper;
+import dev.abarmin.bots.service.support.BotHelper;
+import dev.abarmin.bots.service.support.BotOperation;
+import dev.abarmin.bots.service.support.MessageSourceHelper;
 import dev.abarmin.bots.service.TelegramChatService;
-import dev.abarmin.bots.rss.digest.DigestBotUpdate;
+import dev.abarmin.bots.model.DigestBotUpdate;
 import dev.abarmin.bots.service.RecommendationService;
 import dev.abarmin.bots.service.SubscriptionService;
-import dev.abarmin.bots.rss.reader.persistence.ArticleSource;
+import dev.abarmin.bots.entity.ArticleSource;
 import lombok.RequiredArgsConstructor;
 import org.apache.commons.lang3.StringUtils;
 import org.springframework.context.ApplicationEventPublisher;
