+++ src/main/java/dev/abarmin/bots/service/impl/TelegramUpdateServiceImpl.java
@@ -3,7 +3,7 @@ package dev.abarmin.bots.service.impl;
 import com.pengrad.telegrambot.model.Chat;
 import com.pengrad.telegrambot.model.Message;
 import com.pengrad.telegrambot.model.Update;
-import dev.abarmin.bots.rss.persistence.TelegramBotUpdate;
+import dev.abarmin.bots.entity.TelegramBotUpdate;
 import dev.abarmin.bots.repository.TelegramBotUpdateRepository;
 import dev.abarmin.bots.service.TelegramChatService;
 import dev.abarmin.bots.service.TelegramUpdateService;
