+++ src/main/java/dev/abarmin/bots/service/support/BotHelper.java
@@ -1,9 +1,9 @@
-package dev.abarmin.bots.core;
+package dev.abarmin.bots.service.support;
 
 import com.pengrad.telegrambot.model.Chat;
 import com.pengrad.telegrambot.model.Message;
 import com.pengrad.telegrambot.model.Update;
-import dev.abarmin.bots.rss.persistence.TelegramBotChat;
+import dev.abarmin.bots.entity.TelegramBotChat;
 import dev.abarmin.bots.service.TelegramChatService;
 import lombok.RequiredArgsConstructor;
 import org.apache.commons.lang3.StringUtils;
