+++ src/main/java/dev/abarmin/bots/service/support/BotOperation.java
@@ -1,4 +1,4 @@
-package dev.abarmin.bots.core;
+package dev.abarmin.bots.service.support;
 
 import com.pengrad.telegrambot.model.Update;
