+++ src/main/java/dev/abarmin/bots/service/support/UriJdbcConverter.java
@@ -1,4 +1,4 @@
-package dev.abarmin.bots.core;
+package dev.abarmin.bots.service.support;
 
 import org.springframework.core.convert.converter.Converter;
 import org.springframework.data.convert.ReadingConverter;
