+++ src/test/java/dev/abarmin/bots/publisher/MessagePublisher.java
@@ -5,7 +5,7 @@ import com.pengrad.telegrambot.model.Chat;
 import com.pengrad.telegrambot.model.Message;
 import com.pengrad.telegrambot.model.Update;
 import com.pengrad.telegrambot.request.SendMessage;
-import dev.abarmin.bots.rss.digest.DigestBotUpdate;
+import dev.abarmin.bots.model.DigestBotUpdate;
 import org.mockito.ArgumentCaptor;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.ApplicationEventPublisher;
